自定义stream.groupby-->bigDecimal累加

+++ src/main/java/test/lambdamap/Item.java
@@ -2,6 +2,8 @@ package test.lambdamap;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 /**
  * Created by deweydu
  * Date on 2019/9/29 14:25
@@ -10,8 +12,9 @@ import lombok.Data;
 public class Item {
     private String name;
     private int qty;
-    private double price;
-    public Item(String name,int qty,double price){
+//    private double price;
+    private BigDecimal price;
+    public Item(String name,int qty,BigDecimal price){
         this.name = name;
         this.qty = qty;
         this.price = price;
